requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
django-mathfilters==1.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
et-xmlfile==1.1.0
gunicorn==20.0.4
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
sweetify==2.0.0
typing-extensions==3.10.0.0
whitenoise==5.2.0
XlsxWriter==3.0.2
